requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.11
djangorestframework==3.17.1
dnspython==2.8.0
email-validator==2.3.0
idna==3.11
kombu==5.6.2
mssql-django==1.6
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
pyodbc==5.3.0
pypdf==6.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
qrcode==8.2
redis==7.4.0
reportlab==4.4.10
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.11.0
